Moves Sample.Debugging project into Samples dir
