--- src/Sample.Debugging/App.config
@@ -1,18 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.0"/>
-    </startup>
-  <runtime>
-    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <dependentAssembly>
-        <assemblyIdentity name="Microsoft.Dynamic" publicKeyToken="7f709c5b713576e1" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-1.1.2.22" newVersion="1.1.2.22" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Microsoft.Scripting" publicKeyToken="7f709c5b713576e1" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-1.1.2.22" newVersion="1.1.2.22" />
-      </dependentAssembly>
-    </assemblyBinding>
-  </runtime>
-</configuration>
