+++ src/Samples/Sample.Debugging/packages.config
